refactor: Native modules
